Use MUnit in PingManagerSuite
